LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-gainnode-interface/no-dezippering-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: InvalidStateError: The object is in an invalid state.
CONSOLE MESSAGE: Unhandled Promise Rejection: InvalidStateError: The object is in an invalid state.

Harness Error (FAIL), message = Unhandled rejection: The object is in an invalid state.

PASS # AUDIT TASK RUNNER STARTED. 
FAIL Executing "test0" promise_test: Unhandled rejection with value: object "ReferenceError: Can't find variable: ConstantSourceNode"
PASS Executing "test2" 
PASS Executing "test3" 
PASS Audit report 
PASS > [test0] Dezippering of GainNode removed 
PASS > [test2] Compare value setter and setValueAtTime 
FAIL X .value setter output expected to be equal to the array [0,0.057564035058021545,0.11493717133998871,0.17192906141281128,0.22835084795951843,0.2840152382850647,0.3387378752231598,0.39233696460723877,0.44463497400283813,0.4954586327075958,0.5446387529373169,0.5920130610466003,0.6374237537384033,0.6807206869125366,0.7217600345611572,0.7604058980941772...] but differs in 640 places:
	Index	Actual			Expected
	[384]	-1.2533240020275116e-1	-3.7599720060825348e-2
	[385]	-1.8223451077938080e-1	-5.4670356214046478e-2
	[386]	-2.3853251338005066e-1	-7.1559756994247437e-2
	[387]	-2.9403924942016602e-1	-8.8211774826049805e-2
	...and 636 more errors. assert_true: expected true got false
FAIL X .value setter output matches setValueAtTime output is not true. Got false. assert_true: expected true got false
FAIL < [test2] 2 out of 2 assertions were failed. assert_true: expected true got false
PASS > [test3] Automation effects 
FAIL X With modulation: .value setter output expected to be equal to the array [0,0.06507764756679535,0.14468508958816528,0.2377234548330307,0.34252625703811646,0.45691272616386414,0.5782617330551147,0.7035987377166748,0.8297000527381897,0.9532026648521423,1.0707192420959473,1.178961157798767,1.2748475074768066,1.3556175231933594,1.418926477432251,1.462929368019104...] but differs in 640 places:
	Index	Actual			Expected
	[384]	-1.2533259391784668e-1	-3.7599910050630569e-2
	[385]	-2.0602117478847504e-1	-7.8457005321979523e-2
	[386]	-3.0026960372924805e-1	-1.3329683244228363e-1
	[387]	-4.0656360983848572e-1	-2.0073612034320831e-1
	...and 636 more errors. assert_true: expected true got false
FAIL X With modulation: .value setter output matches setValueAtTime output is not true. Got false. assert_true: expected true got false
FAIL < [test3] 2 out of 2 assertions were failed. assert_true: expected true got false
FAIL # AUDIT TASK RUNNER FINISHED: 2 out of 3 tasks were failed. assert_true: expected true got false
